Fix unit test bugs
