+++ tajo-catalog/tajo-catalog-server/src/test/java/org/apache/tajo/catalog/TestCatalogAgainstCaseSensitivity.java
@@ -27,6 +27,7 @@ import org.apache.tajo.common.TajoDataTypes.Type;
 import org.apache.tajo.exception.UndefinedPartitionException;
 import org.apache.tajo.exception.UndefinedTableException;
 import org.apache.tajo.util.CommonTestingUtil;
+import org.apache.tajo.util.TUtil;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -39,6 +40,7 @@ import static org.junit.Assert.*;
 public class TestCatalogAgainstCaseSensitivity {
   static CatalogServer server;
   static CatalogService catalog;
+  final static String TEST_URI = "hdfs://xxx.com/warehouse";
 
   @BeforeClass
   public static void setup() throws Exception {
@@ -155,14 +157,18 @@ public class TestCatalogAgainstCaseSensitivity {
     assertTrue(catalog.existsTable("TestDatabase1", "testPartition1"));
 
     String partitionName = "DaTe=bBb/dAtE=AaA";
-    PartitionDesc partitionDesc = CatalogTestingUtil.buildPartitionDesc(partitionName);
+    PartitionDesc partitionDesc = CatalogUtil.buildPartitionDesc(partitionName, TEST_URI);
 
-    AlterTableDesc alterTableDesc = new AlterTableDesc();
-    alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
-    alterTableDesc.setPartitionDesc(partitionDesc);
-    alterTableDesc.setAlterTableType(AlterTableType.ADD_PARTITION);
-
-    catalog.alterTable(alterTableDesc);
+    // Disable the alter table add partition statement temporarily at TAJO-1887
+//    AlterTableDesc alterTableDesc = new AlterTableDesc();
+//    alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
+//    alterTableDesc.setPartitionDesc(partitionDesc);
+//    alterTableDesc.setAlterTableType(AlterTableType.ADD_PARTITION);
+//
+//    catalog.alterTable(alterTableDesc);
+    List targetPartitions = TUtil.newList();
+    targetPartitions.add(CatalogUtil.buildPartitionDescProto(partitionName, TEST_URI));
+    catalog.addPartitions("TestDatabase1", "TestPartition1", targetPartitions, true);
 
     PartitionDescProto resultDesc = catalog.getPartition("TestDatabase1", "TestPartition1",
         partitionName);
@@ -173,14 +179,18 @@ public class TestCatalogAgainstCaseSensitivity {
     assertEquals(resultDesc.getPartitionKeysCount(), 2);
 
     partitionName = "DaTe=BbB/dAtE=aAa";
-    partitionDesc = CatalogTestingUtil.buildPartitionDesc(partitionName);
-
-    alterTableDesc = new AlterTableDesc();
-    alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
-    alterTableDesc.setPartitionDesc(partitionDesc);
-    alterTableDesc.setAlterTableType(AlterTableType.ADD_PARTITION);
-
-    catalog.alterTable(alterTableDesc);
+    // Disable the alter table add partition statement temporarily at TAJO-1887
+//    partitionDesc = CatalogTestingUtil.buildPartitionDesc(partitionName);
+//
+//    alterTableDesc = new AlterTableDesc();
+//    alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
+//    alterTableDesc.setPartitionDesc(partitionDesc);
+//    alterTableDesc.setAlterTableType(AlterTableType.ADD_PARTITION);
+//
+//    catalog.alterTable(alterTableDesc);
+    targetPartitions.clear();
+    targetPartitions.add(CatalogUtil.buildPartitionDescProto(partitionName, TEST_URI));
+    catalog.addPartitions("TestDatabase1", "TestPartition1", targetPartitions, true);
 
     resultDesc = catalog.getPartition("TestDatabase1", "TestPartition1",
         partitionName);
@@ -221,9 +231,9 @@ public class TestCatalogAgainstCaseSensitivity {
     //////////////////////////////////////////////////////////////////////////////
 
     partitionName = "DaTe=BbB/dAtE=aAa";
-    partitionDesc = CatalogTestingUtil.buildPartitionDesc(partitionName);
+    partitionDesc = CatalogUtil.buildPartitionDesc(partitionName, TEST_URI);
 
-    alterTableDesc = new AlterTableDesc();
+    AlterTableDesc alterTableDesc = new AlterTableDesc();
     alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
     alterTableDesc.setPartitionDesc(partitionDesc);
     alterTableDesc.setAlterTableType(AlterTableType.DROP_PARTITION);
@@ -237,7 +247,7 @@ public class TestCatalogAgainstCaseSensitivity {
     assertNull(resultDesc);
 
     partitionName = "DaTe=bBb/dAtE=AaA";
-    partitionDesc = CatalogTestingUtil.buildPartitionDesc(partitionName);
+    partitionDesc = CatalogUtil.buildPartitionDesc(partitionName, TEST_URI);
 
     alterTableDesc = new AlterTableDesc();
     alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
