+++ tajo-catalog/tajo-catalog-server/src/test/java/org/apache/tajo/catalog/TestCatalogExceptions.java
@@ -31,16 +31,19 @@ import org.apache.tajo.common.TajoDataTypes.Type;
 import org.apache.tajo.exception.*;
 import org.apache.tajo.util.CommonTestingUtil;
 import org.apache.tajo.util.KeyValueSet;
+import org.apache.tajo.util.TUtil;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
 import java.net.URI;
+import java.util.List;
 
 public class TestCatalogExceptions {
 
   static CatalogServer server;
   static CatalogService catalog;
+  final static String TEST_URI = "hdfs://xxx.com/warehouse";
 
   @BeforeClass
   public static void setup() throws Exception {
@@ -161,43 +164,55 @@ public class TestCatalogExceptions {
   public void testAddPartitionWithWrongUri() throws Exception {
     // TODO: currently, wrong uri does not occur any exception.
     String partitionName = "DaTe=/=AaA";
-    PartitionDesc partitionDesc = CatalogTestingUtil.buildPartitionDesc(partitionName);
-
-    AlterTableDesc alterTableDesc = new AlterTableDesc();
-    alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
-    alterTableDesc.setPartitionDesc(partitionDesc);
-    alterTableDesc.setAlterTableType(AlterTableType.ADD_PARTITION);
-
-    catalog.alterTable(alterTableDesc);
+    // Disable the alter table add partition statement temporarily at TAJO-1887
+//    PartitionDesc partitionDesc = CatalogUtil.buildPartitionDesc(partitionName, "");
+//
+//    AlterTableDesc alterTableDesc = new AlterTableDesc();
+//    alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
+//    alterTableDesc.setPartitionDesc(partitionDesc);
+//    alterTableDesc.setAlterTableType(AlterTableType.ADD_PARTITION);
+//
+//    catalog.alterTable(alterTableDesc);
+    List targetPartitions = TUtil.newList();
+    targetPartitions.add(CatalogUtil.buildPartitionDescProto(partitionName, TEST_URI));
+    catalog.addPartitions("TestDatabase1", "TestPartition1", targetPartitions, true);
   }
 
-  @Test(expected = DuplicatePartitionException.class)
+  // Disable the alter table add partition statement temporarily at TAJO-1887
+//  @Test(expected = DuplicatePartitionException.class)
+  @Test(expected = TajoInternalError.class)
   public void testAddDuplicatePartition() throws Exception {
     String partitionName = "DaTe=bBb/dAtE=AaA";
-    PartitionDesc partitionDesc = CatalogTestingUtil.buildPartitionDesc(partitionName);
-
-    AlterTableDesc alterTableDesc = new AlterTableDesc();
-    alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
-    alterTableDesc.setPartitionDesc(partitionDesc);
-    alterTableDesc.setAlterTableType(AlterTableType.ADD_PARTITION);
-
-    catalog.alterTable(alterTableDesc);
+    // Disable the alter table add partition statement temporarily at TAJO-1887
+//    PartitionDesc partitionDesc = CatalogUtil.buildPartitionDesc(partitionName, TEST_URI);
+//    AlterTableDesc alterTableDesc = new AlterTableDesc();
+//    alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
+//    alterTableDesc.setPartitionDesc(partitionDesc);
+//    alterTableDesc.setAlterTableType(AlterTableType.ADD_PARTITION);
+//
+//    catalog.alterTable(alterTableDesc);
+    List targetPartitions = TUtil.newList();
+    targetPartitions.add(CatalogUtil.buildPartitionDescProto(partitionName, TEST_URI));
+    catalog.addPartitions("TestDatabase1", "TestPartition1", targetPartitions, true);
 
     partitionName = "DaTe=bBb/dAtE=AaA";
-    partitionDesc = CatalogTestingUtil.buildPartitionDesc(partitionName);
-
-    alterTableDesc = new AlterTableDesc();
-    alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
-    alterTableDesc.setPartitionDesc(partitionDesc);
-    alterTableDesc.setAlterTableType(AlterTableType.ADD_PARTITION);
-
-    catalog.alterTable(alterTableDesc);
+    // Disable the alter table add partition statement temporarily at TAJO-1887
+//    partitionDesc = CatalogUtil.buildPartitionDesc(partitionName, TEST_URI);
+//
+//    alterTableDesc = new AlterTableDesc();
+//    alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
+//    alterTableDesc.setPartitionDesc(partitionDesc);
+//    alterTableDesc.setAlterTableType(AlterTableType.ADD_PARTITION);
+//
+//    catalog.alterTable(alterTableDesc);
+    targetPartitions.add(CatalogUtil.buildPartitionDescProto(partitionName, TEST_URI));
+    catalog.addPartitions("TestDatabase1", "TestPartition1", targetPartitions, true);
   }
 
   @Test(expected = UndefinedTableException.class)
   public void testAddPartitionToUndefinedTable() throws Exception {
     String partitionName = "DaTe=bBb/dAtE=AaA";
-    PartitionDesc partitionDesc = CatalogTestingUtil.buildPartitionDesc(partitionName);
+    PartitionDesc partitionDesc = CatalogUtil.buildPartitionDesc(partitionName, TEST_URI);
 
     AlterTableDesc alterTableDesc = new AlterTableDesc();
     alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "undefined"));
@@ -210,7 +225,7 @@ public class TestCatalogExceptions {
   @Test(expected = UndefinedPartitionException.class)
   public void testDropUndefinedPartition() throws Exception {
     String partitionName = "DaTe=undefined/dAtE=undefined";
-    PartitionDesc partitionDesc = CatalogTestingUtil.buildPartitionDesc(partitionName);
+    PartitionDesc partitionDesc = CatalogUtil.buildPartitionDesc(partitionName, TEST_URI);
 
     AlterTableDesc alterTableDesc = new AlterTableDesc();
     alterTableDesc.setTableName(CatalogUtil.buildFQName("TestDatabase1", "TestPartition1"));
